Removed the admin model and merged it with the user model
Admin model seemed redundant as no extra information was saved on the
admin and by using a isAdmin field in the user modal we can easily check
if the user is admin or not and then grant the request
