+++ src/models/user.model.ts
@@ -6,6 +6,7 @@ const userSchema = new Schema<IUser>({
   username: { type: String, required: true },
   email: { type: String, required: true, unique: true },
   password: { type: String, required: true },
+  isAdmin: { type: Boolean, default: false },
   isEmailVerified: { type: Boolean, required: true },
   branch: { type: String, required: true },
   passingYear: { type: String, required: true },
