+++ src/routes/index.ts
@@ -1,7 +1,3 @@
 import userRoutes from './user.routes';
-import statusRoutes from './status.routes';
 
-export default {
-  userRoutes,
-  statusRoutes,
-};
+export default { userRoutes };
