--- src/routes/status.routes.ts
@@ -1,8 +1,0 @@
-import { Router } from 'express';
-import statusController from '../controller/status.controller';
-
-const router = Router();
-
-router.get('/', statusController.getLoginStatus);
-
-export default router;
