+++ src/routes/user.routes.ts
@@ -5,6 +5,7 @@ import isUserAuth from '../middleware/isUserAuth';
 const router = Router();
 
 router.post('/login', userController.loginUser);
+router.get('/status', userController.getLoginStatus);
 router.post('/register', userController.registerUser);
 router.post('/forgot-password', userController.forgotPassword);
 router.post('/reset-password/:token', userController.resetPassword);
