--- src/services/admin.service.ts
@@ -1,9 +1,0 @@
-import adminModel from '../models/admin.model';
-
-const adminService = {
-  findUser: (email: string) => {
-    return adminModel.findOne({ email });
-  },
-};
-
-export default adminService;
