--- src/types/admin.types.ts
@@ -1,6 +1,0 @@
-export interface IAdmin {
-  username: string;
-  email: string;
-  password: string;
-  isEmailVerified: boolean;
-}
