+++ src/types/user.types.ts
@@ -2,6 +2,7 @@ export interface IUser {
   username: string;
   email: string;
   password: string;
+  isAdmin: boolean;
   isEmailVerified: boolean;
   branch: string;
   passingYear: string;
